requirements.txt
absl-py==1.0.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
arrow==1.2.2
astroid==2.11.5
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.3.0
bleach==5.0.0
boto3==1.23.1
botocore==1.26.1
bravado-core==5.17.0
bravado==11.0.3
brotlipy==0.7.0
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
commonmark==0.9.1
cryptography==36.0.0
cycler==0.11.0
dearpygui==1.6.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.4
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.3
fqdn==1.5.1
frozenlist==1.3.0
fsspec==2022.3.0
future==0.18.2
gitdb==4.0.9
gitpython==3.1.27
google-auth-oauthlib==0.4.6
google-auth==2.6.6
grpcio==1.46.1
humanfriendly==10.0
idna==3.3
imageio-ffmpeg==0.4.7
imageio==2.17.0
imgviz==1.5.0
importlib-metadata==4.11.3
iniconfig==1.1.1
ipykernel==6.13.0
ipython-genutils==0.2.0
ipython==8.3.0
ipywidgets==7.7.0
isoduration==20.11.0
isort==5.10.1
jedi==0.18.1
jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
json5==0.9.8
jsonpointer==2.3
jsonref==0.2
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyter-packaging==0.12.0
jupyter-server==1.17.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
jupyterlab==3.4.2
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
markdown==3.3.7
markupsafe==2.1.1
matplotlib-inline==0.1.3
matplotlib==3.5.1
mccabe==0.7.0
mistune==0.8.4
monotonic==1.6
msgpack==1.0.3
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
neptune-client==0.16.2
nest-asyncio==1.5.5
networkx==2.8
ninja==1.10.2.3
notebook-shim==0.1.0
notebook==6.4.11
numpy==1.23.0
oauthlib==3.2.0
open3d==0.15.2
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.0.1
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pydeprecate==0.3.2
pygments==2.12.0
pyjwt==2.4.0
pylint==2.13.9
pymcubes==0.1.2
pyopenssl==22.0.0
pyparsing==3.0.8
pyquaternion==0.9.9
pyrsistent==0.18.1
pysdf==0.1.8
pysocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
pytorch-lightning==1.6.3
pytz==2022.1
pywavelets==1.3.0
pyyaml==6.0
pyzmq==22.3.0
requests-oauthlib==1.3.1
requests==2.27.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==12.3.0
rsa==4.8
s3transfer==0.5.2
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
send2trash==1.8.0
sentry-sdk==1.5.12
setproctitle==1.2.3
setuptools==61.2.0
shortuuid==1.0.9
simplejson==3.17.6
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
swagger-spec-validator==2.7.4
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.9.0
tensorboardx==2.5
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.10.2
torch-ema==0.3
torch-tb-profiler==0.4.0
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.8.2
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
trimesh==3.11.2
typing-extensions==4.1.1
uri-template==1.2.0
urllib3==1.26.9
wandb==0.12.16
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.3.2
werkzeug==2.1.2
wheel==0.37.1
widgetsnbextension==3.6.0
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0